(broken) Udp outgoing for go?
