+++ mirrord-layer/src/error.rs
@@ -93,6 +93,9 @@ pub(crate) enum LayerError {
     #[error("mirrord-layer: Failed to get `Sender` for sending tcp response!")]
     SendErrorTcpResponse,
 
+    #[error("mirrord-layer: Failed to get `Sender` for sending udp response!")]
+    SendErrorUdpResponse,
+
     #[error("mirrord-layer: JoinError failed with `{0}`!")]
     Join(#[from] tokio::task::JoinError),
