+++ mirrord-layer/src/socket/hooks.rs
@@ -100,12 +100,12 @@ pub(crate) unsafe extern "C" fn listen_detour(sockfd: RawFd, backlog: c_int) -> 
 }
 
 #[hook_guard_fn]
-pub(super) unsafe extern "C" fn connect_detour(
+pub(crate) unsafe extern "C" fn connect_detour(
     sockfd: RawFd,
     raw_address: *const sockaddr,
     address_length: socklen_t,
 ) -> c_int {
-    trace!("connect_detour -> sockfd {:#?}", sockfd);
+    debug!("connect_detour -> sockfd {:#?}", sockfd);
 
     // TODO: Is this conversion safe?
     let address = SockAddr::new(*(raw_address as *const _), address_length);
