+++ mirrord-protocol/src/outgoing/udp.rs
@@ -1,0 +1,16 @@
+use super::*;
+use crate::RemoteResult;
+
+#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
+pub enum LayerUdpOutgoing {
+    Connect(LayerConnect),
+    Write(LayerWrite),
+    Close(LayerClose),
+}
+
+#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
+pub enum DaemonUdpOutgoing {
+    Connect(RemoteResult<DaemonConnect>),
+    Read(RemoteResult<DaemonRead>),
+    Close(ConnectionId),
+}
