+++ mirrord-protocol/src/tcp.rs
@@ -4,8 +4,6 @@ use bincode::{Decode, Encode};
 
 use crate::{ConnectionId, Port};
 
-pub mod outgoing;
-
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct NewTcpConnection {
     pub connection_id: ConnectionId,
