+++ tests/go-e2e-outgoing/main.go
@@ -7,9 +7,13 @@ import (
 	"log"
 	"net/http"
 )
+import (
+	"net"
+	"os"
+)
 
-func main() {
-	fmt.Println(">> starting go")
+func request() {
+	fmt.Println(">> making request")
 
 	res, err := http.Get("http://www.google.com/robots.txt")
 	if err != nil {
@@ -27,5 +31,23 @@ func main() {
 		log.Fatal(err)
 	}
 
-	fmt.Printf("%s", body)
+	fmt.Printf("done")
+}
+
+func lookup() {
+	ips, err := net.LookupIP("google.com")
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Could not get IPs: %v\n", err)
+		os.Exit(1)
+	}
+	for _, ip := range ips {
+		fmt.Printf("google.com. IN A %s\n", ip.String())
+	}
+}
+
+func main() {
+	fmt.Println(">> starting go")
+
+	// request()
+	lookup()
 }
